refactor(models): move QueueEntry from orchestra to models
Move QueueEntry dataclass from vibe3.orchestra.queue_entry to
vibe3.models.queue_entry to establish proper domain-first architecture.

Changes:
- Create src/vibe3/models/queue_entry.py with QueueEntry dataclass
- Add QueueEntry re-export to src/vibe3/models/__init__.py
- Convert src/vibe3/orchestra/queue_entry.py to compatibility shell
- Update orchestra/__init__.py lazy import to use models.queue_entry
- Update domain imports to use vibe3.models.queue_entry
- Remove queue_entry from architecture test allowed imports
- Update test imports to use vibe3.models.queue_entry

This breaks the domain → orchestra dependency for queue data contracts.
Domain layer can now import QueueEntry from models (leaf module) instead
of orchestra, establishing proper architectural boundaries.

Backward compatibility preserved via re-export shell in orchestra.queue_entry.

Issue: #1886
