+++ src/vibe3/models/queue_entry.py
@@ -1,0 +1,12 @@
+"""Queue entry data contract."""
+
+from dataclasses import dataclass
+
+
+@dataclass
+class QueueEntry:
+    """Frozen queue entry tracked only by issue identity and wait state."""
+
+    issue_number: int
+    collected_state: str | None = None
+    waiting_state: str | None = None
