+++ tests/vibe3/orchestra/test_dispatch_orphan_flow_context.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 from unittest.mock import MagicMock, patch
 
 from vibe3.models.orchestration import IssueInfo, IssueState
+from vibe3.models.queue_entry import QueueEntry
 from vibe3.orchestra.global_dispatch_coordinator import GlobalDispatchCoordinator
-from vibe3.orchestra.queue_entry import QueueEntry
 
 if TYPE_CHECKING:
     pass
